source/requirementsQuartWebService.txt
python-dotenv==1.0.0
quart>=0.19.0
aiohttp>=3.8.0
hypercorn==0.14.3 
quart-wtforms>=0.4.0
werkzeug==3.0.0